Web/Whacky Telegram bot v.2.0/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.14.2
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
coverage==7.2.3
dill==0.3.6
exceptiongroup==1.1.0
flake8==6.0.0
frozenlist==1.3.3
humanfriendly==10.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
mypy==1.0.1
mypy-extensions==1.0.0
openai==0.27.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.16.3
pyTelegramBotAPI==4.10.0
pytest==7.2.2
pytest-cov==4.0.0
python-dotenv==1.0.0
requests==2.28.2
telebot==0.0.5
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
types-requests==2.28.11.15
types-urllib3==1.26.25.8
typing_extensions==4.5.0
udpy==2.0.0
urllib3==1.26.14
wrapt==1.15.0
yarl==1.8.2
